+++ KinKal/BField.hh
@@ -12,7 +12,7 @@
 namespace KinKal {
   class BField {
     public:
-      typedef ROOT::Math::SMatrix<float,3> Grad; // field gradient: ie dBi/d(x,y,z)
+      typedef ROOT::Math::SMatrix<double,3> Grad; // field gradient: ie dBi/d(x,y,z)
       // return value of the field at a point
       virtual Vec3 fieldVect(Vec3 const& position) const = 0; 
       // return BField gradient = dB_i/dx_j, at a given point
@@ -24,7 +24,7 @@ namespace KinKal {
       // ( = - integral of the 'external' force needed to keep the particle onto this trajectory);
       template <class KTRAJ> void integrate(KTRAJ const& ktraj, TRange const& range, Vec3& dmom) const;
       // speed of light in units to convert Tesla to mm (bending radius)
-      static float cbar() { static float retval = CLHEP::c_light/1000.0; return retval; }
+      static double cbar() { static double retval = CLHEP::c_light/1000.0; return retval; }
 
       virtual ~BField(){}
   };
@@ -36,7 +36,7 @@ namespace KinKal {
       virtual Grad fieldGrad(Vec3 const& position) const override { return Grad(); }
       virtual Vec3 fieldDeriv(Vec3 const& position, Vec3 const& velocity) const override { return Vec3(); }
       UniformBField(Vec3 const& bnom) : fvec_(bnom) {}
-      UniformBField(float BZ) : UniformBField(Vec3(0.0,0.0,BZ)) {}
+      UniformBField(double BZ) : UniformBField(Vec3(0.0,0.0,BZ)) {}
       virtual ~UniformBField(){}
     private:
       Vec3 fvec_; // constant field
@@ -59,32 +59,32 @@ namespace KinKal {
   // simple Z gradient field, used to test Field corrections
   class GradBField : public BField {
     public:
-      GradBField(float b0, float b1, float zg0, float zg1);
+      GradBField(double b0, double b1, double zg0, double zg1);
       virtual Vec3 fieldVect(Vec3 const& position) const override;
       virtual Grad fieldGrad(Vec3 const& position) const override { return fgrad_; }
       virtual Vec3 fieldDeriv(Vec3 const& position, Vec3 const& velocity) const override;
       virtual ~GradBField(){}
     private:
-      float b0_, b1_;
-      float z0_; 
-      float grad_; // gradient in tesla/mm, computed from the fvec values
+      double b0_, b1_;
+      double z0_; 
+      double grad_; // gradient in tesla/mm, computed from the fvec values
       Grad fgrad_;
   };
 
   template<class KTRAJ> void BField::integrate(KTRAJ const& ktraj, TRange const& trange, Vec3& dmom) const {
     dmom = Vec3();
     // compare this field with the noiminal at a few points. only bother to integrate if this is outside tolerance.
-    std::vector<float> tvals = {trange.low(),trange.mid(),trange.high()};
-    std::vector<float> db;
+    std::vector<double> tvals = {trange.low(),trange.mid(),trange.high()};
+    std::vector<double> db;
     for(auto tval :tvals){
       Vec3 bf = fieldVect(ktraj.position(tval));
       db.push_back( (bf - ktraj.bnom(tval)).R());
     }
     if(*std::max_element(db.begin(),db.end()) > 1e-6){ // tolerance should be a parameter FIXME!
       unsigned nsteps(10); // this should be computed from field rate-of-change over this range FIXME!
-      float dt = trange.range()/float(nsteps);
+      double dt = trange.range()/double(nsteps);
       for(unsigned istep=0; istep< nsteps; istep++){
-	float tstep = trange.low() + istep*dt;
+	double tstep = trange.low() + istep*dt;
 	Vec3 vel = ktraj.velocity(tstep);
 	Vec3 db = fieldVect(ktraj.position(tstep)) - ktraj.bnom(tstep);
 	dmom += cbar()*ktraj.charge()*dt*vel.Cross(db);
