+++ KinKal/D2T.hh
@@ -10,29 +10,29 @@ namespace KinKal {
   class D2T { 
     public:
     // given a drift DOCA and direction, compute mean drift time, local speed, and expected RMS of drift time, and the local drift speed
-      virtual void distanceToTime(Pol2 const& drift, float& tdrift, float& tdriftvar, float& dspeed) const = 0;
-      virtual float averageDriftSpeed() const = 0; // average drift speed
-      virtual float maximumDriftTime() const = 0; // maximum drift
+      virtual void distanceToTime(Pol2 const& drift, double& tdrift, double& tdriftvar, double& dspeed) const = 0;
+      virtual double averageDriftSpeed() const = 0; // average drift speed
+      virtual double maximumDriftTime() const = 0; // maximum drift
       virtual ~D2T(){}
   };
 
   // simple implementation of the above using a constant drift velocity and no ExB effects.  Used for testing
   class CVD2T  : public D2T {
     public:
-      virtual void distanceToTime(Pol2 const& drift, float& tdrift, float& tdriftvar, float& dspeed) const override {
+      virtual void distanceToTime(Pol2 const& drift, double& tdrift, double& tdriftvar, double& dspeed) const override {
 	tdrift  = drift.R()/dvel_;
 	tdriftvar = tvar_; 
 	dspeed = dvel_;
       }
       // provide seed (mm/ns) and time RMS (ns) on construction
-      CVD2T(float s, float tvar, float rcell) :dvel_(s), tvar_(tvar), rcell_(rcell) {}
-      virtual float averageDriftSpeed() const override { return dvel_; }
-      virtual float maximumDriftTime() const override { return rcell_/dvel_; }
+      CVD2T(double s, double tvar, double rcell) :dvel_(s), tvar_(tvar), rcell_(rcell) {}
+      virtual double averageDriftSpeed() const override { return dvel_; }
+      virtual double maximumDriftTime() const override { return rcell_/dvel_; }
       virtual ~CVD2T(){}
     private:
-      float dvel_; // constant drift speed
-      float tvar_; // constant time variance
-      float rcell_; // cell radius, used to compute maximum drift
+      double dvel_; // constant drift speed
+      double tvar_; // constant time variance
+      double rcell_; // cell radius, used to compute maximum drift
   };
 
 }
